New interface: ObjectDeserializable
As we're going to let workspaces to configure a lot of things
through a JSON document, this interface offers a method to
deserialize a generic PHP object into an instance of the current
object.

To implement the interface, the following method have to be implemented:

public static function loadFromObject ($data);

$data will typically come from a call to json_decode.

It should return an object of the current class.