Refactoring: new controller ErrorPageController
[ New method to show a 404 page ]

ErrorPageController::show($context, 404);

If you wish to show other HTTP code error pages, update the code to
handle this code and consider to create a new view.

[ Previous workflow ]

If an error 404 should be returned, the code defined ERROR_PAGE to 404,
then included the procedural controllers/errorpage.php file.

[ New workflow ]

If an  error 404 should be returned, a call to ErrorPageController::show
is made.