Fixed bug: Language::findLanguage didn't always return language
[ Previous state ]

When the intersection between the languages accepted by the browser
and the languages we offer exists, but this is not the first language
wanted by the user, it didn't returned this language.

[ Current state ]

The first common language wanted by the browser configuratino, and
offered by the server is well returned.

[ Upstream ]

Note: this change fixes a Keruald bug, reported upstream at
http://bugzilla.espace-win.org/19