I18n :: Language::getConfigVar() now triggers E_USER_NOTICE
Language::getConfigVar ($key) is the private method to get the
value of the Smarty configuration variable $key. This is used to
get the localized string $key.

When the key isn't found, it returns instead #$key# (e.g. #foo#).
This is a behavior added in Azhàr.

Now, it also triggers an error at E_USER_NOTICE level:
The l10n key 'foo' doesn't exist.