Improving error handler code, throwExceptionErrorHandler
[ Feature descriptions ]

We provide a first error handler, throwExceptionErrorHandler to transform
the non recoverable errors into exceptions.

It handles all errors except the four following types:
* E_NOTICE
* E_USER_NOTICE
* E_DEPRECATED
* E_USER_DEPRECATED

[ Future plans ]

This is a transition code, as in the roadmap, for the error handling epic,
we want to offer logging through Apache log4php and through a modern errors
reporting service. We'll output nice 500 pages with feedback invite too.