Move everything to workspaces/
Summary:
Repository root contained directly web application content.

Documentation lives in another repository.

To prepare the application to be more modular, this change puts the application
in a core directory, "workspaces", with src/ and tests/ subfolder.

Paths in tests are updated, even if plans are to switch to PSR-4 autoloading.