Update CSS framework to Bootstrap 3.4.1
CSS was using Bootstrap 3.0.3 from 2014.
Upgrade to last 3.x release, 3.4.1, from 2019.

To upgrade to Bootstrap 5 breaks the navigation and requires more work,
so in a first phase, we limit the update by staying in 3.x.

From there, we'll be able to explore other options, as:
  - updating navigation to work with Bootstrap 5
  - switch to another framework
  - refresh the whole interface

Also fixes CSS issues noticed during upgrade:
  - dl can benefit from more space between blocks
  - we can reduce vertical scrollbar using vh
  - avoid #wrapper and <nav> to use more width than the viewport

Differential Revision: https://devcentral.nasqueron.org/D3828